refactor(utils): rename constants files
