+++ src/lib/utils/media-queries.js
@@ -1,7 +1,7 @@
 import { css as emotionCSS } from '@emotion/core';
 import facepaint from 'facepaint';
 
-import { BreakpointsAll } from 'constants/breakpoints';
+import { BreakpointsAll } from 'constants/Breakpoints';
 
 const mediaQueries = BreakpointsAll.map((bp) => `@media(min-width: ${bp}px)`);
 const mq = facepaint(mediaQueries);
